new updates

+++ lib/models/movie.dart
@@ -39,7 +39,7 @@ class Movie {
         originalLanguage: json['original_language'],
         originalTitle: json['original_title'],
         overview: json['overview'],
-        popularity: json['popularity'],
+        popularity: json['popularity'].toDouble(),
         posterPath: json['poster_path'],
         releaseDate: json['release_date'],
         title: json['title'],
